Update setup-RedHat.yml to support for Centos 9
Hi, I'm kind of new to Ansible so sorry if this is butchering your nice module ;)
I needed Centos 9 support, more a Debian guy to be honest.
Google told me I need the epel-release repo, what ever that is, to be able to install certbot package.
So as always "this works for me", hope this helps.
Btw. I never understood why register a... variable like `dnf_module_enable` and "do not use it", because I cannot find any other reference of it.